Inline and decouple test code from bene_tests.py and contract_tests.py into v*/regression_suite.py; remove bene_tests.py and contract_tests.py
